# docs/api/api-spec.yaml
openapi: 3.0.3
info:
  title: Six Perfections Assessment API
  description: |
    षट्पारमिता (Six Perfections) Assessment API provides endpoints for measuring and developing 
    the six Buddhist perfections (paramitas) through scientifically validated assessments and 
    personalized training programs.
    
    ## The Six Perfections (परमिता)
    
    1. **Generosity (दान - Dana)** - The perfection of giving and selfless sharing
    2. **Ethics (शील - Sila)** - The perfection of moral conduct and ethical behavior  
    3. **Patience (क्षान्ति - Ksanti)** - The perfection of patience and emotional regulation
    4. **Energy (वीर्य - Virya)** - The perfection of spiritual energy and perseverance
    5. **Meditation (ध्यान - Dhyana)** - The perfection of concentration and mindfulness
    6. **Wisdom (प्रज्ञा - Prajna)** - The perfection of understanding and insight
    
    ## Authentication
    
    Most endpoints require JWT authentication via the Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```
    
    ## Rate Limiting
    
    API requests are limited to 100 requests per 15-minute window per IP address.
    
  version: 1.0.0
  contact:
    name: Six Perfections Development Team
    email: api@sixperfections.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.sixperfections.app/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  minLength: 8
                  example: "securePassword123"
                firstName:
                  type: string
                  example: "Tenzin"
                lastName:
                  type: string
                  example: "Norbu"
                languagePreference:
                  type: string
                  enum: [en, ru, zh, ja, bo, sa]
                  default: en
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate user login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  # User management endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT authentication token
        refreshToken:
          type: string
          description: Refresh token for token renewal
        user:
          $ref: '#/components/schemas/User'
        expiresAt:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [user, premium_user, researcher, teacher, admin]
        spiritualLevel:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...